Cleaning up a little to move things to the proper location
