+++ earth_enterprise/src/common/khMTTypes.h
@@ -25,6 +25,7 @@
 #include <set>
 #include <unordered_map>
 #include <mutex>
+#include <vector>
 
 template <class T>
 class khAtomic {
@@ -219,6 +220,63 @@ class khMTMap
     // todo: this is only used for setting up an empty dirtymap by AssetD and AssetVersionD
     // should probably really remove this
   }
+};
+
+// WARNING: Use this with extreme caution
+// It protects some obvious problems but inheriting from vector publicly isn't great
+template <typename T>
+class MTVector : public std::vector<T> {
+  private:
+    mutable std::mutex mtx;
+
+    MTVector(const MTVector& a, const std::lock_guard<std::mutex> &) : Base(a) { }
+  public:
+    typedef std::vector<T> Base;
+    typedef typename Base::const_iterator const_iterator;
+    void push_back(const T& v) {
+      std::lock_guard<std::mutex> lock(mtx);
+      Base::push_back(v);
+    }
+    MTVector() : Base() { }
+    // forward to private copy constructor to protect vector from modification
+    MTVector(const MTVector& a) : MTVector(a, std::lock_guard<std::mutex>(a.mtx)) { }
+    const T& operator[] (const size_t idx) const {
+      std::lock_guard<std::mutex> lock(mtx);
+      return Base::operator[](idx); 
+    }
+
+    bool operator==(const MTVector& x) const {
+      std::lock_guard<std::mutex> lock(mtx);
+      return true;
+    }
+
+    bool operator==(const Base& x) const {
+      std::lock_guard<std::mutex> lock(mtx);
+      return true;
+    }
+
+    void clear(void) {
+      std::lock_guard<std::mutex> lock(mtx);
+      Base::clear();
+    }
+
+    void shrink_to_fit(void) {
+      std::lock_guard<std::mutex> lock(mtx);
+      Base::shrink_to_fit();
+    }
+
+    const_iterator begin(void) const {
+      return Base::begin();
+    }
+
+    const_iterator end(void) const {
+      return Base::end();
+    }
+
+    size_t size(void) const {
+      std::lock_guard<std::mutex> lock(mtx);
+      return Base::size();
+    }
 };
 
 class khTargetedLock {
