+++ earth_enterprise/src/fusion/autoingest/MiscConfigStorage.idl
@@ -35,8 +35,6 @@ class MiscConfigStorage {
   uint        MutexTimedWaitSec = uint(30);
 
   uint        BuildIndexThreads = uint(1);
-
-  uint        ContLoadAndThink = uint(0);
 #pragma LoadAndSave
 
 #hquote
