+++ earth_enterprise/src/fusion/autoingest/plugins.src/RasterProject.src
@@ -1042,7 +1042,7 @@ uint32 RasterProjectAssetVersionImplD::BuildIndex(
   int threadsToUse = MiscConfig::Instance().BuildIndexThreads;
   #pragma omp parallel for ordered schedule(static,1) num_threads(threadsToUse)
   for (uint i = 0; i < genInfo.size(); ++i) {
-    notify(NFY_WARN, "Starting loop %d from thread %d", i+1, omp_get_thread_num());
+    notify(NFY_DEBUG, "Starting loop %d from thread %d", i+1, omp_get_thread_num());
     // Check if I even need to do any packet generation at all
     // Maybe the other insets will completely replace everything
     // that this one would generate
@@ -1168,7 +1168,6 @@ uint32 RasterProjectAssetVersionImplD::BuildIndex(
   timeDiff = std::chrono::duration<double>(curTime - prevTime);
   notify(NFY_INFO2, "BuildIndex finished looping through insets - %f seconds elapsed",
          timeDiff.count());
-  assert(MiscConfig::Instance().ContLoadAndThink);
 
   // Make index subasset (RasterGEIndex).
   return MakeIndexSubAsset(packgenkids, packgenassets, packgenvers, genInfo, date);
